requirements.txt
flask_socketio
flask 
pymongo[srv] 
requests
python-dotenv 